requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.7
django-bootstrap3==7.0.1
kombu==3.0.35
Pillow==3.2.0
pytz==2016.4
redis==2.10.5
